+++ lib/crowd_crush_web.ex
@@ -56,8 +56,10 @@ defmodule CrowdCrushWeb do
       use Phoenix.Channel
 
       import CrowdCrushWeb.{ErrorHelpers, Gettext}
+      import CrowdCrushWeb.UserSocket, except: [connect: 2, id: 1]
 
       alias Phoenix.View
+      alias CrowdCrush.Simulation
       alias CrowdCrushWeb.Endpoint
 
       require Logger