+++ lib/crowd_crush_web/channels/user_socket.ex
@@ -1,4 +1,7 @@
 defmodule CrowdCrushWeb.UserSocket do
+  @moduledoc """
+  User Socket
+  """
 
   use Phoenix.Socket
 
@@ -18,4 +21,15 @@ defmodule CrowdCrushWeb.UserSocket do
   end
 
   def id(socket), do: "users_socket:#{socket.assigns.user_id}"
+
+  @doc """
+  Prints the current time (UTC) in a ISO 8601 formatted string.
+
+  ## Examples
+
+    CrowdCrushWeb.UserSocket.now()
+    "2019-06-17T22:00:49.878834"
+
+  """
+  def now, do: NaiveDateTime.to_iso8601(NaiveDateTime.utc_now())
 end